separate console styles into another file
